+++ oidc-rs-spring-boot-common/src/main/java/org/entur/auth/spring/common/server/JWKSourceWithIssuer.java
@@ -1,16 +1,19 @@
-package org.entur.auth.spring.config.server;
+package org.entur.auth.spring.common.server;
 
 import com.nimbusds.jose.KeySourceException;
 import com.nimbusds.jose.jwk.JWK;
 import com.nimbusds.jose.jwk.JWKSelector;
+import com.nimbusds.jose.jwk.source.JWKSetBasedJWKSource;
+import com.nimbusds.jose.jwk.source.JWKSetSource;
 import com.nimbusds.jose.jwk.source.JWKSource;
 import com.nimbusds.jose.proc.SecurityContext;
 import java.util.List;
 import lombok.Getter;
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public class JWKSourceWithIssuer implements JWKSource<SecurityContext> {
+public class JWKSourceWithIssuer<C extends SecurityContext> implements JWKSource<C> {
     @Getter private final String issuerUrl;
 
     private final JWKSource<SecurityContext> jwkSource;
@@ -23,5 +26,13 @@ public class JWKSourceWithIssuer implements JWKSource<SecurityContext> {
 
     public boolean getReadiness() {
         return true;
+    }
+
+    public @NonNull JWKSetSource<SecurityContext> getJWKSetSource() {
+        if (jwkSource instanceof JWKSetBasedJWKSource<SecurityContext> source)
+            return source.getJWKSetSource();
+
+        throw new IllegalStateException(
+                "Not an instance of %s".formatted(JWKSetBasedJWKSource.class.getSimpleName()));
     }
 }