+++ oidc-rs-spring-boot-web-config/src/main/java/org/entur/auth/spring/config/ConfigAuthManagerResolverAutoConfiguration.java
@@ -3,7 +3,6 @@ package org.entur.auth.spring.config;
 import com.nimbusds.jose.jwk.source.JWKSetSourceWithHealthStatusReporting;
 import com.nimbusds.jose.proc.SecurityContext;
 import com.nimbusds.jose.util.health.HealthReportListener;
-import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -12,10 +11,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.entur.auth.spring.common.server.AuthProviders;
 import org.entur.auth.spring.common.server.EnturAuthProperties;
+import org.entur.auth.spring.common.server.JWKSourceWithIssuer;
 import org.entur.auth.spring.common.server.ServerCondition;
 import org.entur.auth.spring.common.server.TenantJwtGrantedAuthoritiesConverter;
 import org.entur.auth.spring.config.server.IssuerAuthenticationManagerResolver;
-import org.entur.auth.spring.config.server.JWKSourceWithIssuer;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
@@ -37,7 +36,7 @@ import org.springframework.security.authentication.AuthenticationManagerResolver
 @RequiredArgsConstructor
 public class ConfigAuthManagerResolverAutoConfiguration {
     private final Map<String, AuthenticationManager> authenticationManagers = new HashMap<>();
-    private final List<JWKSourceWithIssuer> remoteJWKSets = new ArrayList<>();
+    private final List<JWKSourceWithIssuer<?>> remoteJWKSets = new ArrayList<>();
 
     private final EnturAuthProperties enturAuthProperties;
     private final AuthProviders authProviders;
@@ -47,7 +46,7 @@ public class ConfigAuthManagerResolverAutoConfiguration {
             healthReportListener;
 
     @Bean
-    public AuthenticationManagerResolver<HttpServletRequest> authenticationManagerResolver() {
+    public IssuerAuthenticationManagerResolver authenticationManagerResolver() {
         log.debug("Configure AuthenticationManagerResolver");
         final var authoritiesConverter = new TenantJwtGrantedAuthoritiesConverter(authProviders);